API/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
httpx==0.26.0
python-dotenv==1.0.1
python-dateutil==2.8.2
requests==2.31.0
biopython==1.83
redis==5.0.1
aiohttp==3.9.3
async-timeout==4.0.3 